+++ expyfun/_externals/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from .decorator import decorator  # noqa
+from . import h5io