+++ expyfun/io/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
 from ._wav import read_wav, write_wav
-from ._hdf5 import read_hdf5, write_hdf5
 from ._parse import read_tab